Refactor headers to separate components.
