+++ client/src/components/About.js
@@ -1,15 +1,14 @@
 import React, { Component } from "react";
-import LeftMenu from "./components/LeftMenu";
-import Header from "./components/Header";
+// import LeftMenu from "./LeftMenu";
+import Header from "./Header";
 
 class About extends Component {
   render() {
     return (
       <div className="container container--body">
-        <Header />
+        <Header headerCopy={this.props.header} />
       </div>
     );
   }
 }
 export default About;
-// TODO: Refactor out the headers to be reusable.
