+++ client/src/components/Algos.js
@@ -1,22 +1,13 @@
 import React, { Component } from "react";
+import Header from "./Header";
 
 class Algos extends Component {
   render() {
     return (
       <div className="container container--body">
-        <div className="header">
-          <h2 className="header__title">Algorithyms</h2>
-          <div className="header__copy">
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris
-              cursus purus quis pharetra aliquet. Orci varius natoque penatibus
-              et magnis.
-            </p>
-          </div>
-        </div>
+        <Header headerCopy={this.props.header} />
       </div>
     );
   }
 }
 export default Algos;
-// TODO: Refactor out the headers to be reusable.
